pyproject.toml
[tool.poetry]
name = "LLM Bot"
version = "0.1"
description = "Chat Bot for answering questions about company details"
readme = "README.md"
authors = [
    "Andrii Hladkyi <andrii.hladkyi.teacher@gmail.com>",
]
license = ""

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.10.23"
llama-index-embeddings-openai = "^0.1.7"
llama-index-vector-stores-chroma = "^0.1.6"
llama-index-vector-stores-milvus = "^0.1.6"
fastapi = "^0.110.0"
jinja2 = "^3.1.3"
scipy = "^1.12.0"
pymilvus = "^2.4.0"
flake8-pyproject = "^1.2.3"
pdfminer-six = "^20231228"
python-multipart = "^0.0.9"
llama-index-llms-huggingface = "^0.1.4"
flagembedding = {git = "https://github.com/FlagOpen/FlagEmbedding.git"}
llama-index-postprocessor-flag-embedding-reranker = "^0.1.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
